Fixed networking 2/2, also added packets

+++ src/main/java/com/carbon/thaumicmastery/core/Utils.java
@@ -8,7 +8,7 @@ public class Utils {
 		return player.getEntityData().getBoolean(LibMisc.TAG_SHIELD);
 	}
 
-	public static int dist(int x, int y, int z, int xx, int yy, int zz) {
-		return (int) Math.sqrt((Math.pow(Math.abs(xx - x), 2) + Math.pow(Math.abs(yy - y), 2) + Math.pow(Math.abs(zz - z), 2)));
+	public static double dist(int x, int y, int z, int xx, int yy, int zz) {
+		return Math.sqrt((Math.pow(Math.abs(xx - x), 2) + Math.pow(Math.abs(yy - y), 2) + Math.pow(Math.abs(zz - z), 2)));
 	}
 }
